Remove fd functions, re-standardise PS1
Investigated Solaris issues for PS1 and tput a bit more, figured out a workaround.  This allows the PS1 vars to be much cleaner.

Also built in an auto-switch between a short form of the prompt when the terminal window is smaller, and the full form when the window is wider.

TODO: figure out a switch to prevent PS1 updates e.g. 'no-ps1()'