Added git function
All attempts that I have seen so far to get git information into a PS1 require making a git call every time the prompt is updated.  This has a performance hit, especially for non-git usage and is inefficient.  Every time git is run, however, is an opportunity to grab the latest state of things and shove them into environment variables.  I only care about the current branch for now.