Evict pathfind() and grow get_command()
Evict pathfind() to gists as I never use it.

The search for a specific sudo binary path was from when I worked in an environment with mixed Solaris and Linux, and on the Solaris hosts, sudo could be in any of a number of locations.  This made pre-ansible fleet scripting quite insufferable, so this test and report was a way of organically ensuring that a standard was applied fleetwide.  I haven't worked in that environment in a while now, and see little point in keeping that test/report code around.

The /bin/bash location check is still sane IMHO, and I've expanded get_command() to cover multiple commands and return 1 on any failures (similar to a requires() function) as well as to allow a verbose output.  This verbose output allows get_command() to fully replace pathfind().