Adjust git() to handle main vs master
Repos are defaulting to 'main' now rather than 'master'.  If we run something like 'git checkout master' in such a repo, we get a vague and useless output.  This adjustment detects any usage of the word 'master', checks if the current repo has a 'main' branch, and if this is the case, it suggests options.

We must be very careful to NOT automatically and blindly swap 'master' with 'main'.  That would be easy to do, but also very dangerous.